Whatever will be, will be 8{

+++ bootstrap.ps1
@@ -204,10 +204,17 @@ process
         New-Item -Path $bsPath -ItemType Directory -Force | Out-Null
     }
 
-    $logCommand = "Import-Csv (Get-ChildItem -Path $bsPath\*.csv | Sort-Object -Property LastWriteTime -Descending)[0].FullName | Format-Table Timestamp, Message -AutoSize"
     $logScriptFilePath = "$bsPath\log.ps1"
-    Invoke-ExpressionWithLogging -command "New-Item -Path $logScriptFilePath -ItemType File -Force | Out-Null"
-    Invoke-ExpressionWithLogging -command "Set-Content -Value `"$logCommand`" -Path $logScriptFilePath -Force"
+    if (Test-Path -Path $logScriptFilePath -PathType Leaf)
+    {
+        Write-PSFMessage "$logScriptFilePath already exists, don't need to create it"
+    }
+    else
+    {
+        $logCommand = "Import-Csv (Get-ChildItem -Path $bsPath\*.csv | Sort-Object -Property LastWriteTime -Descending)[0].FullName | Format-Table Timestamp, Message -AutoSize"
+        Invoke-ExpressionWithLogging -command "New-Item -Path $logScriptFilePath -ItemType File -Force | Out-Null"
+        Invoke-ExpressionWithLogging -command "Set-Content -Value `"$logCommand`" -Path $logScriptFilePath -Force"
+    }
 
     if (Get-Module -Name Defender -ListAvailable -ErrorAction SilentlyContinue)
     {
